requirements.txt
openai>=1.0.0
anthropic>=0.18.0
groq>=0.4.0
python-dotenv>=1.0.0
tqdm>=4.65.0
gspread>=5.12.0
google-auth>=2.22.0
google-auth-oauthlib>=1.1.0
